requirement.txt
selenium
requests
colorama
webdriver-manager
beautifulsoup4
azure-storage-blob
azure-identity
msgraph-sdk
msal
httpx
python-dotenv
pillow